use the str() method of cpp_dec_float for serialization
Use the str() method of the cpp_dec_float for serializing the
object into a string. Use the constuctor to deserialize the string
back into a cpp_dec_float object.
With this, use the existing read and write functions for k2 String
to implement the read and write functions for the cpp_dec_float
types.