Using boost archive along with customized streambuf for cpp_dec_float types
All tests passed, but need to revise approach for skvhttpclient MPackSerialization.

need to override the sputc, xsputn, xbumpc, and xsgetn functions of
the std::streambuf class in the customized streambuf class.

using flags = boost::archive::no_header | boost::archive::no_tracking
when creating the boost archive can ensure correct handling of serialization
and deserialization.

On k2 payload side, the customized streambuf works directly with payload content.
Thus, only one copy is involved. The customized streambuf can invoke functions
of the payload to do both byte-wise and block write and read. So, this is also
efficient.

However, on the skvhttpclient MPackSerialization, the mpack write doesn't allow
writing a binary without knowing the size in advance. This prevents block write
(and read). This commits implements the xsputn and xsgetn by treating the buffer
as an array of bytes/chars. This could by quite inefficient.